Add support for [ProvidesFromLib] section
This section is by default enabled.  You can disable it by adding
"enable = false" or "disable = true".  When enabled, it will create the
"provides" section of the META6.json from the files in "lib".  If disabled,
it will assume the author is responsible for the "provides" section of
the META6.json and won't touch it.