Tighten dependencies to latest/greatest and pin fez version
The latest version of fez has a slightly different API which *may*
cause issues with App::Mi6.  It therefore felt safer to pin it to
the last version prior to version 100 until it has been determined
that there are no issues with the latest fez version.

Tightened up other dependencies, because that's a good thing to do
in a world where accountability is becoming more important.

Also update copyright year and add tags