Account for changes in .fez-config.json fil
It appears that the newer version of fez has changed the format of
the .fez-config file with regards to groups, which made it impossible
for people to do an upload for e.g. Raku Community modules.

Instead of:

    "groups": [
      {
        "group": "raku-community-modules",

the new format uses "name" as the key, instead of "group":

    "groups": [
      {
        "name": "raku-community-modules",

This commit fixes that by first checking for the new key ("name"),
and if that didn't yield a result, try the old key ("group").

This also removes a warning about using undefined values for anybody
that has a fez group membership of any kind.